train.py
"""
ASA-DETR训练脚本
"""
import argparse
import os
import yaml
import torch
import torch.nn as nn
from torch.utils.data import DataLoader
from pathlib import Path

# 导入模型和工具（需要根据实际实现调整）
# from models.asa_detr import ASADETR
# from utils.dataset import LandslideDataset
# from utils.loss import ASADETRLoss
# from utils.trainer import Trainer


def parse_args():
    """解析命令行参数"""
    parser = argparse.ArgumentParser(description='ASA-DETR训练脚本')
    parser.add_argument('--cfg', type=str, default='configs/asa-detr.yaml',
                        help='配置文件路径')
    parser.add_argument('--data', type=str, required=True,
                        help='数据集配置文件路径')
    parser.add_argument('--weights', type=str, default='',
                        help='预训练权重路径')
    parser.add_argument('--epochs', type=int, default=150,
                        help='训练轮数')
    parser.add_argument('--batch-size', type=int, default=8,
                        help='批次大小')
    parser.add_argument('--device', type=str, default='0',
                        help='GPU设备ID')
    parser.add_argument('--project', type=str, default='runs/train',
                        help='保存路径')
    parser.add_argument('--name', type=str, default='exp',
                        help='实验名称')
    return parser.parse_args()


def load_config(cfg_path):
    """加载配置文件"""
    with open(cfg_path, 'r', encoding='utf-8') as f:
        cfg = yaml.safe_load(f)
    return cfg


def main():
    """主函数"""
    args = parse_args()
    
    # 加载配置
    cfg = load_config(args.cfg)
    
    # 设置设备
    device = torch.device(f'cuda:{args.device}' if torch.cuda.is_available() else 'cpu')
    print(f'使用设备: {device}')
    
    # 创建保存目录
    save_dir = Path(args.project) / args.name
    save_dir.mkdir(parents=True, exist_ok=True)
    
    # TODO: 初始化模型
    # model = ASADETR(cfg['model'])
    # model = model.to(device)
    
    # TODO: 加载预训练权重
    # if args.weights:
    #     checkpoint = torch.load(args.weights, map_location=device)
    #     model.load_state_dict(checkpoint['model'])
    #     print(f'加载预训练权重: {args.weights}')
    
    # TODO: 准备数据集
    # train_dataset = LandslideDataset(...)
    # train_loader = DataLoader(train_dataset, batch_size=args.batch_size, ...)
    
    # TODO: 初始化优化器和学习率调度器
    # optimizer = torch.optim.AdamW(model.parameters(), ...)
    # scheduler = ...
    
    # TODO: 初始化损失函数
    # criterion = ASADETRLoss(cfg['train']['loss'])
    
    # TODO: 开始训练
    # trainer = Trainer(model, train_loader, val_loader, optimizer, 
    #                   scheduler, criterion, device, save_dir)
    # trainer.train(args.epochs)
    
    print('训练脚本模板已创建，请根据实际模型实现补充代码')
    print(f'配置文件: {args.cfg}')
    print(f'数据集: {args.data}')
    print(f'训练轮数: {args.epochs}')
    print(f'批次大小: {args.batch_size}')
    print(f'保存路径: {save_dir}')


if __name__ == '__main__':
    main()