requirements.txt
# 核心依赖
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.3.0
pyyaml>=5.4.0
scipy>=1.7.0
tqdm>=4.62.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.6.0

# 模型分析
thop>=0.1.0
fvcore>=0.1.5

# 数据处理
albumentations>=1.1.0
pycocotools>=2.0.4

# 其他工具
pandas>=1.3.0
scikit-learn>=0.24.0